+++ python/interfaces/BaseManipulation.py
@@ -107,12 +107,12 @@ class BaseManipulation:
         """
         cmd += ' '
         if goal is not None:
-            cmd += 'goal ' + ' '.join(str(f) for f in goal) + ' '
+            cmd += 'goal ' + ' '.join('%.15e'%f for f in goal) + ' '
         if goals is not None:
             cmd += 'goals %d '%len(goals)
             for g in goals:
                 for f in g:
-                    cmd += str(f) + ' '
+                    cmd += '%.15e '%f
         if steplength is not None:
             cmd += 'steplength %.15e '%steplength
         if execute is not None:
@@ -188,7 +188,7 @@ class BaseManipulation:
         """See :ref:`module-basemanipulation-moveunsyncjoints`
         """
         assert(len(jointinds)==len(jointvalues) and len(jointinds)>0)
-        cmd = 'MoveUnsyncJoints handjoints %d %s %s '%(len(jointinds),' '.join(str(f) for f in jointvalues), ' '.join(str(f) for f in jointinds))
+        cmd = 'MoveUnsyncJoints handjoints %d %s %s '%(len(jointinds),' '.join('%.15e'%f for f in jointvalues), ' '.join(str(f) for f in jointinds))
         if planner is not None:
             cmd += 'planner %s '%planner
         if execute is not None:
@@ -240,7 +240,7 @@ class BaseManipulation:
         """
         cmd = 'FindIKWithFilters ikparam %s '%str(ikparam)
         if cone is not None:
-            cmd += 'cone %s '%(' '.join(str(f) for f in cone))
+            cmd += 'cone %s '%(' '.join('%.15e'%f for f in cone))
         if solveall is not None and solveall:
             cmd += 'solveall '
         if filteroptions is not None: